+++ packages/vscode/CONTRIBUTING.md
@@ -18,7 +18,10 @@ On push to the master branch, a new Insider extension is released, with an incre
 ├── packages
 │   └── vscode
 │       └── src
-|           └── extension.ts // Language Client entry point
+|           └── extension.ts // VS Code Entry Point
+│           └── plugins
+│               └── nextjs                  // Next JS entry point
+│               └── prisma-language-server  // Language Client entry point
 |   └── language-server      // Language Server
 │       └── src
 │           └── cli.ts    // Language Server CLI entry point
