+++ packages/vscode/package.json
@@ -36,7 +36,7 @@
   },
   "main": "./dist/src/extension.js",
   "activationEvents": [
-    "onLanguage:prisma",
+    "workspaceContains:**/*.prisma",
     "onCommand:prisma.plugin.nextjs.addTypes"
   ],
   "contributes": {
@@ -76,7 +76,7 @@
     ],
     "configuration": {
       "type": "object",
-      "title": "Prisma Configuration",
+      "title": "Prisma",
       "properties": {
         "prismaLanguageServer.prismaFmtBinPath": {
           "scope": "resource",
@@ -87,6 +87,11 @@
           "scope": "resource",
           "type": "string",
           "description": "Option to set the path to a nextjs project"
+        },
+        "prisma.plugin.nextjs.hasPrompted": {
+          "scope": "resource",
+          "type": "boolean",
+          "description": "Hides Autotypes Prompt if it has been shown"
         },
         "prisma.plugin.nextjs.addTypesOnSave": {
           "scope": "resource",
@@ -99,7 +104,8 @@
     "commands": [
       {
         "command": "prisma.restartLanguageServer",
-        "title": "Prisma: Restart Language Server"
+        "title": "Restart Language Server",
+        "category": "Prisma"
       },
       {
         "command": "prisma.plugin.nextjs.addTypes",
