+++ packages/vscode/src/helpers/workspace.ts
@@ -1,17 +1,40 @@
 import path from 'path'
 import { workspace } from 'vscode'
+type PackageScopes = 'devDependencies' | 'dependencies' | 'peerDependencies'
 export async function packageJsonIncludes(
   packageName: string,
+  scopes: PackageScopes[],
 ): Promise<boolean> {
   const rootPath = workspace.rootPath
   if (rootPath) {
     const packageJson = await require(path.join(rootPath, 'package.json')) // eslint-disable-line
-    // eslint-disable-next-line
-    const dependencies = packageJson && packageJson['dependencies']
-    if (dependencies) {
-      const packages = Object.keys(dependencies)
-      return packages.includes(packageName)
+    for (const scope of scopes) {
+      // eslint-disable-next-line
+      const dependencies = packageJson && packageJson[scope]
+      if (dependencies) {
+        const packages = Object.keys(dependencies)
+        return packages.includes(packageName)
+      }
     }
+  }
+  return false
+}
+
+export async function packageSearch(
+  packageName: string,
+  scopes: PackageScopes[],
+): Promise<boolean> {
+  const rootPath = workspace.rootPath
+  if (rootPath) {
+    const packageJson = await require(path.join(rootPath, 'package.json')) // eslint-disable-line
+    scopes.forEach((scope) => {
+      // eslint-disable-next-line
+      const dependencies = packageJson && packageJson[scope]
+      if (dependencies) {
+        const packages = Object.keys(dependencies)
+        return packages.includes(packageName)
+      }
+    })
   }
   return false
 }
