+++ packages/vscode/src/plugins/types.ts
@@ -1,14 +1,11 @@
 import { ExtensionContext } from 'vscode'
 
-type PluginCommand = {
-  id: string
-  action: (context: ExtensionContext) => Promise<void> | void
-}
 export interface PrismaVSCodePlugin {
   name: string
-  commands?: PluginCommand[]
+  /** This is called during vscodes' activate event and if true will call the plugins activate function */
   enabled: () => Promise<boolean> | boolean
 
+  /** Called when the extension is activated and if enabled returns true */
   activate?: (context: ExtensionContext) => Promise<void> | void
   deactivate?: () => Promise<void> | void
 }
